requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
supabase==2.3.0
joblib==1.3.2
